% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBGsymbols-package.r
\docType{package}
\name{BBGsymbols-package}
\alias{BBGsymbols}
\alias{BBGsymbols-package}
\title{BBGsymbols: Popular Bloomberg tickers and field symbols conveniently packaged for R users}
\description{
A collection of helper datasets for working with Bloomberg financial data in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bautheac.github.oi/BBGsymbols}
  \item Report bugs at \url{https://github.com/bautheac/BBGsymbols/issues}
}

}
\author{
\strong{Maintainer}: Oivier Bauthéac \email{o.bautheac@gmail.com}

}
\keyword{internal}
