% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fields}
\alias{fields}
\title{field symbols for Bloomberg tickers}
\format{A tibble with 5 variables:
\itemize{
  \item{\code{instrument}: underlying financial instrument type (futures, equity, etc.).}
  \item{\code{type}: type of data; instrument specific.
    \itemize{
      \item{futures: market, aggregate}
      \item{equity: market, balance sheet, cash flow statement, income statement, ratios.}
    }
  }
  \item{name: field name.}
  \item{symbol: corresponding Bloomberg field symbol.}
  \item{description: Bloomberg provided description for the corresponding field.}
}}
\source{
Bloomberg
}
\usage{
fields
}
\description{
Helper dataset for querying data from Bloomberg.
  Provides field symbols for various security type-data type combinations
}
\details{
These symbols are for use querying data from Bloomberg.
}
\seealso{
\itemize{
  \item{function \code{\link[pullit]{futures_ticker}} in package \code{pullit}.}
}
}
\keyword{datasets}
