% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{futures}
\alias{futures}
\title{Bloomberg tickers for various popular futures series}
\format{A tibble with 109 rows and 7 variables:
\itemize{
  \item{active contract ticker: active contract ticker for the futures series.}
  \item{asset class: asset class for the underlying futures series; 'commodity', 'currency', 'financial' and 'index'.}
  \item{sector: underlying instrument sector; asset class specific.
    \itemize{
      \item{commodity: 'agriculturals', 'energy', 'metals'.}
      \item{currency: domestic currency.}
      \item{financial: instrument type.}
      \item{index: index asset class.}
    }
  }
  \item{subsector: underlying instrument subsector; asset class specific.
    \itemize{
      \item{commodity:
        \itemize{
          \item{agriculturals: 'grains', 'livestock', 'softs'.}
          \item{energy: 'gas', 'petroleum'.}
          \item{metals: 'base', 'precious'.}
        }
      }
      \item{currency: foreign currency.}
      \item{financial: for government debt underlyings, corresponding country code (ISO 3166-1 alpha-2).}
      \item{index. sector specific:
        \itemize{
          \item{commodity indexes (commodity): commodity sector covered by the index.}
          \item{equity indexes (equity): country/region covered by the index.}
        }
      }
    }
  }
  \item{exchange MIC: ISO 10383 Codes for exchanges and market identification (MIC). Identifies the exchange where the corresponding futures series trades.}
  \item{name: corresponding futures series's underlying name.}
}}
\source{
Bloomberg.
}
\usage{
futures
}
\description{
The dataset includes Bloomberg active contract tickers for several popular futures series including
commodity, currency, financial and index futures with underlyings from various asset classes.
}
\details{
These tickers are for use in querying data on a Bloomberg terminal.
}
\keyword{datasets}
