% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{months}
\alias{months}
\title{Bloomberg futures delivery month symbols}
\format{A data.table. Columns include:
\itemize{
  \item{name: month name.}
  \item{symbol: corresponding symbol.}
}}
\source{
\url{https://www.cmegroup.com/month-codes.html?redirect=/product-codes-listing/month-codes.html}
}
\usage{
months
}
\description{
Lists the conventional letter codes used to specify futures delivery month.
}
\seealso{
\code{pullit} and \code{storethat} packages from the \code{finRes} universe.
}
\keyword{datasets}
