% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rolls}
\alias{rolls}
\title{Roll symbols for futures term structure Bloomberg ticker construction}
\format{A \linkS4class{data.table}. Columns include:
\itemize{
  \item{\code{roll}: divides variables between roll type ('type') and roll adjustment (adjustment) variables.}
  \item{\code{symbol}: roll type or adjustment symbol.}
  \item{\code{name}: corresponding roll type or adjustment name.}
}}
\source{
Bloomberg
}
\usage{
rolls
}
\description{
Helper dataset for Bloomberg futures term structure ticker construction.
  Provides roll type and roll adjustment symbols to use along days and months offsets in
  constructing Bloomberg futures term structure tickers.
}
\seealso{
\itemize{
  \item{function \code{\link[pullit]{futures_ticker}} in package \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}.}
  \item{"GFUT <GO>" &/or "DOCS #2072138 <GO>" on a Bloomberg terminal.}
  \item{The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite in general, in particular the
    \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} and
    \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} packages.}
}
}
\keyword{datasets}
