% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{months}
\alias{months}
\title{Futures delivery month symbols}
\format{A \linkS4class{data.table}. Columns include:
\itemize{
  \item{name: month name.}
  \item{symbol: corresponding symbol.}
}}
\source{
\url{https://www.cmegroup.com/month-codes.html?redirect=/product-codes-listing/month-codes.html}
}
\usage{
months
}
\description{
Lists the Street convention symbols for futures delivery month.
}
\seealso{
\itemize{
    \item{The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite, in particular the
    \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} and
    \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} packages.}
  }
}
\keyword{datasets}
