% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tickers_futures}
\alias{tickers_futures}
\title{Bloomberg tickers for various popular futures series}
\format{A data.table. Columns include:
\itemize{
  \item{\code{ticker}: active contract ticker for the futures series.}
  \item{\code{name}: corresponding futures series's underlying name.}
  \item{\code{asset class}: asset class for the underlying futures series; 'commodity', 'currency', 'financial' and 'index'.}
  \item{\code{sector}: underlying instrument sector; asset class specific.
    \itemize{
      \item{commodity: 'agriculturals', 'energy', 'metals'.}
      \item{currency: domestic currency.}
      \item{financial: instrument type.}
      \item{index: index asset class.}
    }
  }
  \item{\code{subsector}: underlying instrument subsector; asset class specific.
    \itemize{
      \item{commodity:
        \itemize{
          \item{agriculturals: 'grains', 'livestock', 'softs'.}
          \item{energy: 'gas', 'petroleum'.}
          \item{metals: 'base', 'precious'.}
        }
      }
      \item{currency: foreign currency.}
      \item{financial: for government debt underlyings, corresponding country code (ISO 3166-1 alpha-2).}
      \item{index. sector specific:
        \itemize{
          \item{commodity indexes (commodity): commodity sector covered by the index.}
          \item{equity indexes (equity): country/region covered by the index.}
        }
      }
    }
  }
  \item{\code{term structure length}: length of the term structure for the futures series (number of contracts).}
  \item{\code{MIC}: ISO 10383 Codes for exchanges and market identification (MIC). Identifies the exchange where the corresponding futures series trades.}
  \item{\code{currency}: currency for observed price variables.}
  \item{\code{FIGI}: Financial Instrument Global Identifier. Twelve character, alphanumeric identifier. The first 2 characters are upper-case consonants
    (including "Y"), the third character is the upper-case "G", characters 4 -11 are any upper-case consonant (including "Y") or integer between 0 and 9,
    and the last character is a check-digit. An identifier is assigned to instruments of all asset classes, is unique to an individual instrument and once
    issued will not change for an instrument. For equity instruments an identifier is issued per instrument per trading venue.}
}}
\source{
Bloomberg
}
\usage{
tickers_futures
}
\description{
The dataset includes Bloomberg active contract tickers for several popular futures series including
commodity, currency, financial and index futures with underlyings from various asset classes.
}
\details{
These tickers are for use in querying data on a Bloomberg terminal.
}
\seealso{
\code{pullit} and \code{storethat} packages from the \code{finRes} universe.
}
\keyword{datasets}
